[PLAT-1362] Remove ShippingMethods and update deuna now widget
