+++ etc/webapi.xml
@@ -13,12 +13,6 @@
             <resource ref="anonymous"/>
         </resources>
     </route>
-    <route url="/V1/orders/:cartId/shipping-methods" method="POST">
-        <service class="Deuna\Now\Api\ShippingMethodsInterface" method="get"/>
-        <resources>
-            <resource ref="anonymous"/>
-        </resources>
-    </route>
     <route url="/V1/deuna/public-key" method="GET">
         <service class="Deuna\Now\Model\ConfigManagement" method="getPublicKey"/>
         <resources>
